Allow different logical and physical sector size for ZVOL
This patch addresses the issue of Windows showing same logical
and physical sector size for a zvol. If a zvol is created with
different volblocksize (other than 512 and 4K) Windows cannot
handle it properly. This patch makes the zvol's logical sector
size always 512 and physical sector size set to volblocksize.