Fixed wrong assertion in libspl
problem: When the hostid value is not yet defined in the registry the library throws an assertion sometimes.

reason: the value type is not initialized in the stack so when the registry call fails to get the hostid the type value is still compared to REG_DWORD.

fix: do not check for possible assertion on the value type if the call failed.