SSV-19147 - BSOD in ZFSin due to high IRQL
Problem: the pool driver can sometimes issue I/Os through the DI at DISPATCH level from a DPC routine.

Reason: That causes taskq queuing to BSOD because of a mutex used in the taskq mechanism.

Fix: Return to using a workitem to queue a request that will then queue the I/O in the taskq.  Do it for all I/Os, even when running at a level < DISPATCH because we want to return control to the caller as soon as possible and do not want to suffer being preempted by the wait on the taskq mutex.
also aligned the workitem block on a 16-bytes boundary: this is preferable as it is the standard alignment for dynamic memory allocations on a x64 system.

TESTED: Verified on my rig.  QA will re-rerun the TP closedown tests and that will be enough.