Performance enhancements. - mutex handling to use fast-mutex and provide fairness . - added the MUTEX_SPIN type allocation to the spl-mutex class so spinlocks can be provisioned using the same kmutex_t structure. - condvar to not set a condition for setting the event so none can be lost. - taskq enhanced to allow for queueing at dispatch level - reworked the storport and DI routines to no longer use work items but straight queue into taskq. - Enhanced uio to allow for preallocated uio and iovec blocks so no memory allocation occurs there for each I/O (for DI the WrkParm structure is still dynamically allocated). - Added thread priority processing in spl-thread but kept all threads at their system default of 8 (for now, so we do not skew perf results. We'll play with thread priority next).
USE_TASKQ_ZVOL : the module can be compiled with or without using taskq to queue front end I/Os. If not using taskq then workitems are used.  By default taskq is used, workitems are not used.

USE_MUTEX_EVENT : The module can be compiled for spl-mutex to be using existing event mechanism or new fast-mutex mechanism.  By default the new fast-mutex mechanism is used, event is not used.

USE_QUEUED_SPINLOCK : queued spinlocks improve fairness in lock contention. However during my tests it seemed to lock up so for now they are disabled and the standard spinlock is used.