Renamed preprocessor definitions
USE_WORKITEMS_FE : the module can be compiled with or without using taskq to queue front end I/Os. If not using taskq then workitems are used.  By default taskq is used, workitems are not used for front-end I/Os.
USE_MUTEX_EVENT : The module can be compiled for spl-mutex to be using existing event mechanism or new fast-mutex mechanism.  By default the new fast-mutex mechanism is used, event is not used.
USE_QUEUED_SPINLOCK : queued spinlocks improve fairness in lock contention. However during my tests it seemed to lock up so for now they are disabled and the standard spinlock is used.